test: fix test
